Summary figure
This script is for creating and saving a summary plot for intracellular neural data.  Data should be saved as .txt files with each column being a different presentation of the stimulus and each row being a sample. .Txt file can contain header data before the traces begin, but the number of numeric headers must be specified.  This script will batch through all files matching a specified naming structure in a specifies folder and create and save a .tiff image of the summary figure.